// src/types/events.ts
export interface ClobBookEvent {
  tokenId: string
  bids: Array<{ price: number; size: number }>
  asks: Array<{ price: number; size: number }>
  ts: number
}

export interface ClobBestBidAskEvent {
  tokenId: string
  bid: number
  ask: number
  ts: number
}

export interface ClobPriceChangeEvent {
  tokenId: string
  price: number
  size: number
  side: 'BUY' | 'SELL'
  bestBid: number
  bestAsk: number
  ts: number
}

export interface ClobLastTradeEvent {
  tokenId: string
  price: number
  size: number
  side: 'BUY' | 'SELL'
  ts: number
}

export interface ClobTickSizeChangeEvent {
  tokenId: string
  tickSize: number
  ts: number
}